requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
kafka-python==2.0.2
requests==2.32.3
urllib3==2.2.3
